+++ example/lib/special_text/my_special_text_span_builder.dart
@@ -1,15 +1,14 @@
-import 'package:example/special_text/at_text.dart';
-import 'package:example/special_text/dollar_text.dart';
-import 'package:example/special_text/emoji_text.dart';
 import 'package:extended_text_library/extended_text_library.dart';
 import 'package:flutter/material.dart';
+
+import 'at_text.dart';
+import 'dollar_text.dart';
+import 'emoji_text.dart';
 
 class MySpecialTextSpanBuilder extends SpecialTextSpanBuilder {
   /// whether show background for @somebody
   final bool showAtBackground;
-  final BuilderType type;
-  MySpecialTextSpanBuilder(
-      {this.showAtBackground: false, this.type: BuilderType.extendedText});
+  MySpecialTextSpanBuilder({this.showAtBackground: false});
 
   @override
   TextSpan build(String data, {TextStyle textStyle, onTap}) {
@@ -24,18 +23,18 @@ class MySpecialTextSpanBuilder extends SpecialTextSpanBuilder {
 
     ///index is end index of start flag, so text start index should be index-(flag.length-1)
     if (isStart(flag, AtText.flag)) {
-      return AtText(textStyle, onTap,
-          start: index - (AtText.flag.length - 1),
-          showAtBackground: showAtBackground,
-          type: type);
+      return AtText(
+        textStyle,
+        onTap,
+        start: index - (AtText.flag.length - 1),
+        showAtBackground: showAtBackground,
+      );
     } else if (isStart(flag, EmojiText.flag)) {
       return EmojiText(textStyle, start: index - (EmojiText.flag.length - 1));
     } else if (isStart(flag, DollarText.flag)) {
       return DollarText(textStyle, onTap,
-          start: index - (DollarText.flag.length - 1), type: type);
+          start: index - (DollarText.flag.length - 1));
     }
     return null;
   }
 }
-
-enum BuilderType { extendedText, extendedTextField }